+++ CHANGELOG.md
@@ -32,3 +32,4 @@
   - Try to fallback thumbnail generation for video with horribly broken encoding
   - Country blocklist now can actually fit all countries
   - Make "auth level" text box into "account type" dropdown in accounts page, easier to understand
+  - Board owners can now edit custom pages